engine_traits: Move import methods to their own trait
