+++ components/engine_traits/src/lib.rs
@@ -28,6 +28,8 @@ mod cf_handle;
 pub use crate::cf_handle::*;
 mod cf_options;
 pub use crate::cf_options::*;
+mod import;
+pub use import::*;
 
 pub const DATA_KEY_PREFIX_LEN: usize = 1;
