+++ components/engine_traits/src/options.rs
@@ -153,7 +153,3 @@ impl Default for IterOptions {
         }
     }
 }
-
-pub struct IngestExternalFileOptions {
-    pub move_files: bool,
-}
